requirements.txt
python-dotenv==0.10.3
python-telegram-bot==11.1.0
redis==3.4.1
